db/requirements.txt
pytest
ipython
psycopg2-binary
